[DO NOT REVIEW] rename allocator to manager
